+++ Application/Features/Expenses/GetExpenses/Models/GetExpensesRequest.cs
@@ -1,5 +1,5 @@
 ﻿namespace IOU1.Application.Features.Expenses.GetExpenses.Models;
 
-public sealed record GetExpensesRequest(long GroupId)
+public sealed record GetExpensesRequest(int GroupId)
 {
 }
