+++ Application/Features/Groups/AddGroup/Models/AddedGroup.cs
@@ -1,3 +1,3 @@
 ﻿namespace IOU1.Application.Features.Groups.AddGroup.Models;
 
-public record AddedGroup(long GroupId);
+public record AddedGroup(int GroupId);
