+++ Application/Features/Groups/DeleteGroup/Request/DeleteGroupRequest.cs
@@ -4,5 +4,5 @@ namespace IOU1.Application.Features.Groups.DeleteGroup.Request;
 
 public sealed record DeleteGroupRequest : IRequest
 {
-    public required long GroupId { get; init; }
+    public required int GroupId { get; init; }
 }
