+++ Application/Features/Groups/GetGroup/Models/Dto/GetGroupExpenseDto.cs
@@ -2,9 +2,9 @@ namespace IOU1.Application.Features.Groups.GetGroup.Models.Dto;
 
 public sealed record GetGroupExpenseDto
 {
-    public required long ExpenseId { get; init; }
+    public required int ExpenseId { get; init; }
 
-    public required long BuyerId { get; init; }
+    public required int BuyerId { get; init; }
 
     public required decimal Amount { get; init; }
