+++ Application/Features/Groups/GetGroup/Models/Dto/GetGroupExpenseShareDto.cs
@@ -2,9 +2,9 @@ namespace IOU1.Application.Features.Groups.GetGroup.Models.Dto;
 
 public sealed record GetGroupExpenseShareDto
 {
-    public required long ExpenseShareId { get; init; }
+    public required int ExpenseShareId { get; init; }
 
-    public required long MemberId { get; init; }
+    public required int MemberId { get; init; }
 
     public required decimal Amount { get; init; }
