+++ Application/Features/Groups/GetGroup/Query/IGetGroupQuery.cs
@@ -4,7 +4,7 @@ namespace IOU1.Application.Features.Groups.GetGroup.Query;
 
 public interface IGetGroupQuery
 {
-    Task<GetGroupDto?> GetGroup(long groupId, CancellationToken cancellationToken = default);
+    Task<GetGroupDto?> GetGroup(int groupId, CancellationToken cancellationToken = default);
 
-    Task<IEnumerable<GetGroupExpenseDto>> GetLastExpenses(long groupId, int lastExpenseCount, CancellationToken cancellationToken = default);
+    Task<IEnumerable<GetGroupExpenseDto>> GetLastExpenses(int groupId, int lastExpenseCount, CancellationToken cancellationToken = default);
 }
