+++ Application/Features/Groups/GetGroupSummary/Models/Request/GetGroupSummaryRequest.cs
@@ -2,5 +2,5 @@ namespace IOU1.Application.Features.Groups.GetGroupSummary.Models.Request;
 
 public sealed record GetGroupSummaryRequest
 {
-    public required long GroupId { get; init; }
+    public required int GroupId { get; init; }
 }
