+++ Application/Features/Groups/GetGroups/Models/Dto/GetGroupsDto.cs
@@ -2,7 +2,7 @@
 
 public record GetGroupsDto
 {
-    public required long GroupId { get; init; }
+    public required int GroupId { get; init; }
 
     public required string OwnerName { get; init; }
