+++ Application/Features/Invitations/UseInvitationLink/Models/UseInvitationLinkRequest.cs
@@ -4,7 +4,7 @@ namespace IOU1.Application.Features.Invitations.UseInvitationLink.Models;
 
 public record UseInvitationLinkRequest : IRequest
 {
-    public required long? InvitationId { get; init; }
+    public required int? InvitationId { get; init; }
 
     public required string? InvitationKey { get; init; }
 }
