+++ Application/Features/Members/AddMember/Models/AddMemberDto.cs
@@ -1,6 +1,6 @@
 ﻿namespace IOU1.Application.Features.Members.AddMember.Models;
 
 public record AddMemberDto(
-    long UserId,
-    long GroupId,
-    long MemberId);
+    int UserId,
+    int GroupId,
+    int MemberId);
