+++ Application/Features/Members/AddMember/Models/AddMemberRequest.cs
@@ -1,3 +1,3 @@
 ﻿namespace IOU1.Application.Features.Members.AddMember.Models;
 
-public sealed record AddMemberRequest(long UserId, long GroupId);
+public sealed record AddMemberRequest(int UserId, int GroupId);
