+++ Application/Features/Settlements/GetSettlementsOverview/Models/GetSettlementOverviewResponse.cs
@@ -2,16 +2,16 @@
 
 public record GetSettlementOverviewResponse
 {
-    public long GroupId { get; init; }
+    public int GroupId { get; init; }
     public IEnumerable<PaymentInfo> Payments { get; } = [];
 }
 
 public record PaymentInfo
 {
-    public long FromMemberId { get; init; }
+    public int FromMemberId { get; init; }
     public required string FromMemberName { get; init; }
 
-    public long ToMemberId { get; init; }
+    public int ToMemberId { get; init; }
     public required string ToMemberName { get; init; }
 
     public decimal Amount { get; init; }
