+++ Application/Features/Users/DeleteUser/Models/Endpoint/DeleteUserRequest.cs
@@ -2,4 +2,4 @@
 
 namespace IOU1.Application.Features.Users.DeleteUser.Models.Endpoint;
 
-public record DeleteUserRequest(long UserId) : IRequest;
+public record DeleteUserRequest(int UserId) : IRequest;
