+++ Application/Services/Groups/IGroupService.cs
@@ -6,7 +6,7 @@ namespace IOU1.Application.Services.Groups;
 
 public interface IGroupService
 {
-    Task<Result<AddedGroup?>> AddGroup(IEnumerable<long> memberIds, long ownerId, string name, string? description, string currencyKey, CancellationToken cancellationToken = default);
+    Task<Result<AddedGroup?>> AddGroup(IEnumerable<int> memberIds, int ownerId, string name, string? description, string currencyKey, CancellationToken cancellationToken = default);
 
-    Task<Result<DeleteGroupDto?>> DeleteGroup(long groupId, CancellationToken cancellationToken = default);
+    Task<Result<DeleteGroupDto?>> DeleteGroup(int groupId, CancellationToken cancellationToken = default);
 }
