+++ Application/Services/Invitations/IDirectInvitationCreationService.cs
@@ -4,7 +4,7 @@ namespace IOU1.Application.Services.Invitations;
 
 public interface IDirectInvitationCreationService
 {
-    Task<Invitation> MakeInvitation(string email, long groupId, long senderId);
-    Task<UserDto?> CheckInvitationPossible(string email, long groupId, long senderId);
+    Task<Invitation> MakeInvitation(string email, int groupId, int senderId);
+    Task<UserDto?> CheckInvitationPossible(string email, int groupId, int senderId);
 
 }
