+++ Application/Services/Members/IMemberService.cs
@@ -5,7 +5,7 @@ namespace IOU1.Application.Services.Members;
 
 public interface IMemberService
 {
-    Task<Result<GroupMember?>> AddMember(long groupId, long memberId, CancellationToken cancellationToken = default);
+    Task<Result<GroupMember?>> AddMember(int groupId, int memberId, CancellationToken cancellationToken = default);
 
-    Task<bool> IsMemberOfGroup(long groupId, long memberId, CancellationToken cancellationToken = default);
+    Task<bool> IsMemberOfGroup(int groupId, int memberId, CancellationToken cancellationToken = default);
 }
