+++ Domain/Models/Split.cs
@@ -2,7 +2,7 @@
 
 public record Split
 {
-    public required long MemberId { get; init; }
+    public required int MemberId { get; init; }
 
     public required decimal Amount { get; init; }
 }
