+++ Domain/RepoInterfaces/IExpenseRepository.cs
@@ -12,8 +12,8 @@ public enum SortingOptions
 public interface IExpenseRepository : IRepository
 {
     Task AddAsync(Expense expense, CancellationToken cancellationToken = default);
-    Task<Expense?> GetByIdAsync(long id, CancellationToken cancellationToken = default);
-    Task<IEnumerable<Expense>> GetByGroupIdAsync(long groupId, CancellationToken cancellationToken = default);
-    IQueryable<Expense> GetByGroupIdQuery(long groupId);
+    Task<Expense?> GetByIdAsync(int id, CancellationToken cancellationToken = default);
+    Task<IEnumerable<Expense>> GetByGroupIdAsync(int groupId, CancellationToken cancellationToken = default);
+    IQueryable<Expense> GetByGroupIdQuery(int groupId);
     Task SaveChangesAsync(CancellationToken cancellationToken = default);
 }
