+++ Domain/RepoInterfaces/IRepository.cs
@@ -15,7 +15,7 @@ public interface IRepository<T> : IRepository where T : Entity
 
     Task<int> SaveChanges(CancellationToken cancellationToken = default);
 
-    Task<T?> GetById(long id, CancellationToken cancellationToken = default);
+    Task<T?> GetById(int id, CancellationToken cancellationToken = default);
 
     Task<T?> GetFirstOrDefault(Expression<Func<T, bool>> filter, QueryComposer<T>? toBeIncluded = null, Func<IQueryable<T>, IOrderedQueryable<T>>? orderBy = null, bool shouldBeTracked = true, CancellationToken cancellation = default);
 
@@ -23,5 +23,5 @@ public interface IRepository<T> : IRepository where T : Entity
 
     Task<IEnumerable<T>> Get(Expression<Func<T, bool>> filter, QueryComposer<T>? toBeIncluded = null, Func<IQueryable<T>, IOrderedQueryable<T>>? orderBy = null, bool shouldBeTracked = true, CancellationToken cancellation = default);
 
-    Task<T?> Find(long id, CancellationToken cancellationToken = default);
+    Task<T?> Find(int id, CancellationToken cancellationToken = default);
 }
