+++ Infrastructure/Repositories/GroupRepository.cs
@@ -11,7 +11,7 @@ public class GroupRepository(IOU1Context context)
 {
     private readonly IOU1Context _context = context;
 
-    public Task<Group?> GetGroupWithMembers(long groupId, CancellationToken cancellationToken = default)
+    public Task<Group?> GetGroupWithMembers(int groupId, CancellationToken cancellationToken = default)
     {
         return _context.Groups
             .Include(g => g.Members)
@@ -19,12 +19,12 @@ public class GroupRepository(IOU1Context context)
             .SingleOrDefaultAsync(cancellationToken);
     }
 
-    public async Task<Group?> GetByIdAsync(long groupId, CancellationToken cancellationToken = default)
+    public async Task<Group?> GetByIdAsync(int groupId, CancellationToken cancellationToken = default)
     {
         return await _context.Groups.FirstOrDefaultAsync(g => g.Id == groupId, cancellationToken);
     }
 
-    public async Task<Group?> GetGroupWithMembersAsync(long groupId, CancellationToken cancellationToken = default)
+    public async Task<Group?> GetGroupWithMembersAsync(int groupId, CancellationToken cancellationToken = default)
     {
         return await _context.Groups
             .Include(g => g.Members)
